+++ inertia.go
@@ -7,7 +7,6 @@ import (
 	"io/fs"
 	"log"
 	"net/http"
-	"strings"
 )
 
 // Inertia is a main Gonertia structure, which contains all the logic for being an Inertia adapter.
@@ -19,6 +18,9 @@ type Inertia struct {
 	sharedProps         Props
 	sharedTemplateData  TemplateData
 	sharedTemplateFuncs TemplateFuncs
+
+	ssrURL        string
+	ssrHTTPClient *http.Client
 
 	containerID  string
 	version      string
@@ -54,126 +56,4 @@ type marshallJSON func(v any) ([]byte, error)
 type logger interface {
 	Printf(format string, v ...any)
 	Println(v ...any)
-}
-
-// Location creates redirect response.
-//
-// If request was made by Inertia - sets status to 409 and url will be in "X-Inertia-Location" header.
-// Otherwise, it will do an HTTP redirect with specified status (default is 302 for GET, 303 for POST/PUT/PATCH).
-func (i *Inertia) Location(w http.ResponseWriter, r *http.Request, url string, status ...int) {
-	if IsInertiaRequest(r) {
-		setInertiaLocationInResponse(w, url)
-		return
-	}
-
-	redirectResponse(w, r, url, status...)
-}
-
-// Back creates redirect response to the previous url.
-func (i *Inertia) Back(w http.ResponseWriter, r *http.Request, status ...int) {
-	i.Location(w, r, i.backURL(r), status...)
-}
-
-// Render returns response with Inertia data.
-//
-// If request was made by Inertia - it will return data in JSON format.
-// Otherwise, it will return HTML with root template.
-func (i *Inertia) Render(w http.ResponseWriter, r *http.Request, component string, props ...Props) (err error) {
-	page, err := i.buildPage(r, component, firstOr[Props](props, nil))
-	if err != nil {
-		return fmt.Errorf("build page: %w", err)
-	}
-
-	if IsInertiaRequest(r) {
-		if err = i.doInertiaResponse(w, page); err != nil {
-			return fmt.Errorf("inertia response: %w", err)
-		}
-
-		return
-	}
-
-	if err = i.doHTMLResponse(w, r, page); err != nil {
-		return fmt.Errorf("html response: %w", err)
-	}
-
-	return nil
-}
-
-type page struct {
-	Component string `json:"component"`
-	Props     Props  `json:"props"`
-	URL       string `json:"url"`
-	Version   string `json:"version"`
-}
-
-func (i *Inertia) buildPage(r *http.Request, component string, props Props) (*page, error) {
-	props, err := i.prepareProps(r, component, props)
-	if err != nil {
-		return nil, fmt.Errorf("prepare props: %w", err)
-	}
-
-	return &page{
-		Component: component,
-		Props:     props,
-		URL:       r.RequestURI,
-		Version:   i.version,
-	}, nil
-}
-
-func (i *Inertia) doInertiaResponse(w http.ResponseWriter, page *page) error {
-	pageJSON, err := i.marshallJSON(page)
-	if err != nil {
-		return fmt.Errorf("marshal page into json: %w", err)
-	}
-
-	setInertiaInResponse(w)
-	setJSONResponse(w)
-	setResponseStatus(w, http.StatusOK)
-
-	if _, err := w.Write(pageJSON); err != nil {
-		return fmt.Errorf("write bytes to response: %w", err)
-	}
-
-	return nil
-}
-
-func (i *Inertia) doHTMLResponse(w http.ResponseWriter, r *http.Request, page *page) (err error) {
-	// If root template is already created - we'll use it to save some time.
-	if i.rootTemplate == nil {
-		i.rootTemplate, err = i.buildRootTemplate()
-		if err != nil {
-			return fmt.Errorf("build root template: %w", err)
-		}
-	}
-
-	templateData, err := i.buildTemplateData(r, page)
-	if err != nil {
-		return fmt.Errorf("build template data: %w", err)
-	}
-
-	setHTMLResponse(w)
-
-	if err := i.rootTemplate.Execute(w, templateData); err != nil {
-		return fmt.Errorf("execute root template: %w", err)
-	}
-
-	return nil
-}
-
-func (i *Inertia) inertiaContainerHTML(pageJSON []byte) template.HTML {
-	builder := new(strings.Builder)
-
-	// It doesn't look pretty, but fast!
-	builder.WriteString(`<div id="`)
-	builder.WriteString(i.containerID)
-	builder.WriteString(`" data-page="`)
-	template.HTMLEscape(builder, pageJSON)
-	builder.WriteString(`"></div>`)
-
-	return template.HTML(builder.String())
-}
-
-func (i *Inertia) backURL(r *http.Request) string {
-	// At the moment, it based only on the "Referer" HTTP header.
-	return refererFromRequest(r)
 }
