--- validation.go
@@ -1,4 +1,0 @@
-package gonertia
-
-// ValidationErrors are messages, that will be stored in the "errors" prop.
-type ValidationErrors map[string]any
